tests/wpt/tests/html/semantics/embedded-content/media-elements/interfaces/TextTrackCue/id.html

<!doctype html>
<title>TextTrackCue.id</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
setup(function(){
    window.video = document.createElement('video');
    window.t1 = video.addTextTrack('subtitles');
    document.body.appendChild(video);
});
test(function(){
    var c1 = new VTTCue(0, 1, 'text1');
    c1.id = 'id1\r\n\u0000';
    assert_equals(c1.id, 'id1\r\n\u0000');
    c1.id = c1.id;
    assert_equals(c1.id, 'id1\r\n\u0000');
    c1.id = null;
    assert_equals(c1.id, 'null');
}, document.title+', script-created cue');

var t_parsed = async_test(document.title+', parsed cue');
t_parsed.step(function(){
    var t = document.createElement('track');
    t.onload = this.step_func(function(){
        var c = t.track.cues;
        assert_equals(c[0].id, '');
        assert_equals(c[1].id, 'foobar');
        this.done();
    });
    t.onerror = this.step_func(function() {
      assert_unreached('got error event');
    });
    t.src = 'data:text/vtt,'+encodeURIComponent('WEBVTT\n\n00:00:00.000 --> 00:00:00.001\ntest'+
                                                      '\n\nfoobar\n00:00:00.000 --> 00:00:00.001\ntest');
    t.track.mode = 'showing';
    video.appendChild(t);
});
</script>
