LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrackCue/id-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.

PASS TextTrackCue.id, script-created cue 
FAIL TextTrackCue.id, parsed cue assert_unreached: got error event Reached unreachable code
